Updated openstack/openstack Project: openstack/heat  b4a1ad2bd5bcd55b808d9757e41c5eeb30707485
Avoid creating two Stacks when loading Resource

When load()ing a Resource in order to check it, we must load its definition
from whatever version of the template it was created or last updated with.
Previously we created a second Stack object with that template in order to
obtain the resource definition. Since all we really need in order to obtain
this is the StackDefinition, create just that instead.

Change-Id: Ia05983c3d1b838d2e28bb5eca38d13e83ccaf368
Implements: blueprint stack-definition