Updated openstack/openstack Project: openstack-infra/system-config  032654dd4602d33a16dcf294c7e4e0c2880aaa8d
Make apache cache structure saner

Previously we had 5 levels of dirs 3 character names long. This results
in 64 ** 3 dirents possible per dir level which is a ton of dirs. Reduce
it to 2 character names which is 64 ** 2 = 4096 and much more
reasonable. We don't increase the depth becaues this should still
provide more than enough caching ability as is.

Change-Id: Ic498aff57f8cbba679fd1ce4ec4569926bf7655f