Updated openstack/openstack Project: openstack/manila  63f262821f4765d4d933cb13f69124cb35086866
Change the way to create image service

The way of creating service image in Shared File Systems installation
guide is outdated. The parameter 'copy_from' is no longer supported by
Glance V2.This patch mainly for changing the way of uploading an image
to glance.
At the same time, this patch modified the structer of this file to make
it more clearly. Some content is splited and three subtitles are added,
they are: 'Create a share type', 'Create a manila server image', and
'Create a share network'.

Change-Id: Ie21f6c63e41d4de4ef94c21e3a2979943d25489a
Closes-Bug: 1709266