Updated openstack/openstack Project: openstack-dev/grenade  ee538d1e7c1a92cb8412713e03cdfbcdacad7fab
Remove ensure logs check in upgrade scripts

This commit removes the ensure_logs_exist call in all the upgrade
scripts. This function currently only works in a non-systemd world. Now
that all the logs go in the journal we can't check for separate log
files anymore. In the future we'll modifiy the function to handle
systemd and the journal properly, but for right now this is to unblock
things.

Change-Id: Iedf824a1772115e0dff287a898636f8e58471269