Updated openstack/openstack Project: openstack/tripleo-heat-templates  84362d3385f72ec1a1a0d5b792f091c5af6e4afd
Make HA container bundle work on remote nodes

Right now when we deploy an HA bundle on a pacemaker remote node,
the deploy will fail due to the fact that the bundle includes
tripleo::profile::base::pacemaker which makes a call to
hiera('hacluster_pwd') which will fail on pcmk remote nodes.
While we could noop the profile on pcmk nodes, it's much simpler
to just make sure this hiera key exists on pcmk remote nodes.

Also make sure that pacemaker::corosync::manage_fw is set to false
on remote nodes, otherwise the mere inclusion of the pacemaker
profile will cause iptables-save to run in a container and thus failing.

Change-Id: I09b3e54a470cc2d600a701d23463962501c5c9d6