Updated openstack/openstack Project: openstack/diskimage-builder  ea23aa13a5f88955f3043a3ada7d23c0506d0710
Add netbase to ensure /etc/protocols is placed for debian

Many programs rely upon /etc/protocols to be present
however the default debian image that is generated lacks
/etc/protocols. This is observable when building an image
for use with ironic via the ironic-agent element, since
the IPA agent fails to start as python needs /etc/protocols
to open a socket connection.

Added to debian-minimal as it is inherited into the debian
element.

Change-Id: Icc81635870961943707cf6b3f61a9ddbd51cb8fd
Closes-Bug: #1708531