Updated openstack/openstack Project: openstack/tripleo-specs  9d3582c30e6f12ae702067d3e9b25ccd5bb1f9e8
Propose tech-debt policy

We need a formal way to track and be able to report when we do things in
code or during the cycle that should or need to be fixed later. Without
a unified policy to do this, it's hard to really understand where all
the tech-debt lives.  Given the size of tripleo there is already a lot
of existing and undocumented techdebt but there's no reason why we
cannot start tracking it now to make sure we do not continue to add to
the work.

Change-Id: I7d5a7920a14324670d94034a97160189cb7fe891