Updated openstack/openstack Project: openstack/neutron  ba9a308c3f4fac47a18eb9f754fa7eee1fe01b59
tests: don't set description in test_blank_update_clears_association

'description' attribute of floating IPs is part of
standard-attr-description API extension, so using it in the test case
without requiring the extension is incorrect. But instead of adding the
new requirement, this patch removes 'description' from the floating IP
payload because it's not needed to prove the point of the test case
(that an empty update body will disassociate the floating IP).

Change-Id: I9ab5b0ba5f32f73818f8eac8a0001eeb13248423