Updated openstack/openstack Project: openstack-infra/project-config  6bc98b70418eab3aacc007f8b8ba044fa1017332
Skip additional tests for Cinder doc changes

We currently optimize docs or release note only changes so we
don't run long-running tests with no functional code changes.
This adds compliance, bandit, and compliance jobs to the exclusion
list as these fall into the same category of testing things that are
not affected.

Change-Id: I5c1d04d45bca9133b9c902a2c920ad2218c2d800