Updated openstack/openstack Project: openstack-infra/puppet-zuul  a7ef39c2c0df2ca5cebc9cc51b019267d0b925c6
Zuulv3: move the job dir under /var/lib/zuul

We want to ensure that our jobdir and the git repo cache are
on the same filesystem.  Configure the executors to use a dir
under /var/lib/zuul for the jobdir.  We will mount a new (large)
filesystem at /var/lib/zuul to accomodate both.

Change-Id: I91472fa609274aadcf4945aad44f25014c07b105