Updated openstack/openstack Project: openstack-infra/system-config  a07f6d4f8fed1fa47f37ce154d62ae91b04cffce
Revert "Replace buildlogs.centos with buildlogs.cdn.centos"

So it seems that metadata lives on buildlogs.centos.org and RPM on
buildlogs.cdn.centos.org. For the moment, we'll need to support both.
Which means, we still have a 302 redirect hit.

It shouldn't be too bad, because this is how docker.io also works.

This reverts commit d4b73ff77fb44a85209e5cc33ec57d4675042e15.

Change-Id: Id202d41357c42eeb73020a3a6e6c768644ad1139
Signed-off-by: Paul Belanger <pabelanger@redhat.com>