Updated openstack/openstack Project: openstack/keystone  682cfa5c6d135641797ec9e51299287e8191e858
Remove duplicate roles from federated auth

We were using a one-liner to prune duplicate role references from a
list of roles, but it didn't work in all cases. This reworks the
logic to pass the existing test case. I also added a comment
explaining why the logic we used previously doesn't work so we can
hopefully avoid the pattern in the future.

Change-Id: Id786d6463364ad8f4f02c22bb83221baac4b83d0
Closes-Bug: 1701324