Updated openstack/openstack Project: openstack-infra/project-config  eb58a36152f36a578d59e9925cf96a73358b5854
TEMP - Disable voting on tripleo upgrade jobs

We're currentling hitting https://bugs.launchpad.net/tripleo/+bug/1702955
a lot and we're working on it.

At the same time, we need to land
https://review.openstack.org/#/c/489874/ which has been rechecked too
much times now, we need it on master to move forward with upgrades.

This patch is temporary and will be reverted immediately after we land
https://review.openstack.org/#/c/489874/.

Change-Id: I94fc3c17a8320309646d07d3e76879386131e975