Updated openstack/openstack Project: openstack/cinder  28ed0a40ffb406adb7886b559258bb072ef69a22
3PAR: Fixed image cache enabled capability.

This patch fixes the problem with the 3PAR not
working with generic image cache during cache-miss scenario.

In order to do the image caching, the 3PAR has to create a
temporary snapshot of the volume being cloned.  Then the
background cloning is against the temporary snapshot.
The problem is, the temporary snapshot stays around.
We try to account for this temporary snapshot in
the delete volume action.

Addition to above scenario, When we try retype/migration of
above volume then this operation was unsuccessful as
temporary snapshot was present. So before doing retype/
migration of volume on 3par, we are checking temporary
snapshot on it, if it is present then we are deleting it.

Change-Id: I5eae88660b987d30c29a4827d349acc0350148bb
Closes-Bug: 1686985