Updated openstack/openstack Project: openstack-infra/puppet-zuul  43cf9a5990955400f554254d8cfc1c554e56b9e9
Install ara on the executors

zuul v3 supports optionally collecting data via ara if it's installed.
Install it so that we can produce ara html reports for all of our runs.

The current implementation is similar to the devstack-gate
implementation - it uses a sqlite db per job. The bubblewrap invocation
sets $HOME to the work_dir, so the sqlite db for each job goes in to the
root of the work_dir with no further config.

Change-Id: Iefeda37338c57466d671ea07e7d318fa92aafc35
Depends-On: I8facdf0b95b83d43c337058d70fe6bf71e17d570