Updated openstack/openstack Project: openstack-infra/project-config  9092fc78dffcaf31e15b4e2e4d79603217872ebf
Add TripleO scenario007-container experimental job for OVN

Scenario007-container is a new CI job in TripleO that will test containerized OpenStack with
OVN. The job will be turned to voting, once it is stable.

It will only be run in TripleO gate when touching scenario007-container files or
OVN profiles, so it shouldn't take much resources.

Change-Id: If633a52933c1b24c44cfff30b30ffa800a6ac322