Updated openstack/openstack Project: openstack/tripleo-heat-templates  4207a03b4c9866b7d1fb23eca33172592009269b
Remove iscsid from TLS everywhere docker environment

Since nova-compute is not containerized with TLS yet, using containerized
iscsid causes errors when trying to spawn a VM with a volume. Since
the path is different in this case.

I will re-add iscsid to this environment once nova-compute is
containerized with TLS.

bp tls-via-certmonger-containers

Change-Id: Ida87b187e56ae852c5a4ef6f78cc04a0870fe3f4