Updated openstack/openstack Project: openstack/requirements  0f48444a3573a989fec362861f036e247143e062
Add stestr to global requirements

stestr is a fork of testrepository, designed specifically to concentrate
on being a dedicated test runner.[1] Testrepository is used everywhere for
testing in openstack but is not being maintained at all. The last patch
(which wasn't fixing the broken ci config) was on Dec. 6 2016, but it
hasn't been released. I also had several patches up for review that went
years without any progress. [2][3] Having such a critical piece of our
testing infrastructure isn't maintainable situation. stestr was forked
primarily because of the lack of activity in testr, and it fixes a lot
of longstanding bugs and usability issues and is actively maintained. I
also wrote it to try and be simpler to maintain, the testrepository
codebase was hard to contribute to. (at least for me) So I tried to make
the interfaces as straightforward and explicit. The long term plan is to
get everything onto stestr and stop using testrepository for testing in
OpenStack. But in the short term I have patches up to switch os-testr
under the covers [4] and start migrating some projects [5][6] before
making the big push to switch everything.

[Q] Is the library actively maintained?
[A] Yes.

[Q] Is the library good code?
[A] Yes.

[Q] Is the library python 3 compatible?
[A] Yes.

[Q] Is the library license compatible?
[A] Yes. Apache 2.0.

[Q] Is the library already packaged in the distros we target
    (Ubuntu latest / Fedora latest)?
[A] No.

[Q] Is the function of this library already covered by other
    libraries in global-requirements.txt?
[A] No.

[Q] Is the library required for OpenStack project or related dev
    or infrastructure setup? (Answer to this should be Yes,
    of course) Which?
[A] Yes, as described above.

[1] https://github.com/mtreinish/stestr
[2] https://github.com/testing-cabal/testrepository/pull/16
[3] https://github.com/testing-cabal/testrepository/pull/19
[4] https://review.openstack.org/#/c/488441/
[5] https://review.openstack.org/491035
[6] https://review.openstack.org/#/c/491074/

Change-Id: I2c7618a742439fd2ed26879f3114f0f66fd6337f