Updated openstack/openstack Project: openstack-infra/puppet-ansible  65c5bb3d95f935081bf8759a7a7d1ae04f1ed307
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I4f8f403452416a1b6203684e1b6d7b5e2743dce3