Updated openstack/openstack Project: openstack-infra/puppet-iptables  e69236f2e68c4986c9e34641c27daad545216953
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I10a82afb33c487b3914f1f6449e76d7b9e91cf48