Updated openstack/openstack Project: openstack-infra/puppet-logstash  b7032eb94bfe876609d8d85e8bed5323512932ef
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I20bf542923065f7790993e29938c71bfc189e447