Updated openstack/openstack Project: openstack-infra/puppet-elastic_recheck  2261796c78e71cee33cad8ebb46a97321e049032
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I6df94c6dbb8b4198765375619aaf828339a6b6a0