Updated openstack/openstack Project: openstack-infra/puppet-apparmor  80a1265207cc5d1818e6887a026736248c5e24c3
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I7d12c87b38a2798753c9d46d2015837556506d74