Updated openstack/openstack Project: openstack-infra/puppet-drupal  e4901f6217f9fdb4a2818f2706018244f735d23e
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: Iac4b8137c8a8621083ced76879e9f4fbe1ce70a0