Updated openstack/openstack Project: openstack-infra/puppet-jenkins  70840700de562c9f57851880d44fdaf578b5ecf8
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: Icfe2ae1d6379bb55a37be833e974c567595053c5