Updated openstack/openstack Project: openstack-infra/puppet-etherpad_lite  1740cb5c04aacc5ba069ad19688cf8fd2eaa0c19
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I2d36dbc5790dbd97509f52210f313c7e2e5be8df