Updated openstack/openstack Project: openstack-infra/puppet-haveged  4ea93acf4d26b11b16f937721b42c51f01c5816e
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: If0f66d60444c4dbcd0b46772fb017c44bceef1ad