Updated openstack/openstack Project: openstack-infra/puppet-asterisk  b13391ded3dd0ee6a665a2693f6350557911b4d7
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I8b1d3b1c9f1c750978989c705a9718790855dbb7