Updated openstack/openstack Project: openstack-infra/puppet-jeepyb  edceffecfaa8bebfed3a794ea1bc16a2f121ef4a
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: Ib8c5c1bd7d4a5c7175623ed9dc53683d2f592786