Updated openstack/openstack Project: openstack-infra/puppet-github  30bbbfd58f3dc470a92a895fc1391c7de3597848
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: Ib11eae592fda6fc34aa424e8f973cc1f570a30ed