Updated openstack/openstack Project: openstack-infra/puppet-graphite  4a30897c0575c41c03f5d52593d250bb073483f4
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I2c1adf57f6d025d6e7fe9bc1dcb721b9961aac5f