Updated openstack/openstack Project: openstack-infra/puppet-exim  fec74d203f159144ba2f6ac134d3fdd4a73a1285
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: Ic60a4664b47f1f3d37e3dc8831ef620aa128db0a