Updated openstack/openstack Project: openstack-infra/puppet-askbot  7abbdafc622114e009d0ad70f8efae1a4c8c8ae2
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I1bfeb46bb500bff1cefb581b0bf1e2fa7df7b0e8