Updated openstack/openstack Project: openstack-infra/puppet-log_processor  22c22b0356a34a0da476cc5061d57685b29d3112
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I39394696ddb30d4c9ab67c7b3598c7a723c89b14