Updated openstack/openstack Project: openstack-infra/puppet-kibana  65192133b7765b974d5d13e3580da32c8b97c970
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I5b2adaf4f13eed7b7074ddf82dc5cb2da342f8c0