Updated openstack/openstack Project: openstack-infra/puppet-bugdaystats  c7b87455918ed0ec221d4ddedcdc340c2c7a5ecc
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: Ie3fc3263e520b1313825225b4094c8ce923d22db