Updated openstack/openstack Project: openstack-infra/puppet-cgit  e6cac2e1b8af23f2ed7e477185bef4aa10d2eeb3
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: If86144ecaf206ada80aebae350504c0d51495dff