Updated openstack/openstack Project: openstack-infra/puppet-lodgeit  27183a432f70cb2a7a429b13fe4fc93bb6bcd83c
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: Iaf0aa710b252fc61913a95ead55762eee6b31e6e