Updated openstack/openstack Project: openstack-infra/puppet-bup  0b509cf53cc5bfd7ff91c729bc620b8208e65094
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: Ifaeac1bbbf094998e6a513abe548ba90f3461d3e