Updated openstack/openstack Project: openstack-infra/puppet-httpd  efd4267ef4f80e4f37503ee84ae78bb39d6e3483
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I8233333e7c35a3a1191c0f8d070f6de5dfb7b33d