Updated openstack/openstack Project: openstack-infra/puppet-gerritbot  4e969c5808da513048d9e05ac17b51848a923f87
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I1867c3bdd410b2ace3eabee62f3a228bb6ee157c