Updated openstack/openstack Project: openstack-infra/puppet-elasticsearch  32b7f9ee61a3502e7ae85a990ac4691a6186b311
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I090e8ed3c8dbaf5f1a23cf09806d9dbc03c5f143