Updated openstack/openstack Project: openstack-infra/puppet-kerberos  7f3f2aa0740640f67a8f552bfe0e1926166c48e8
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I2dea0aa652ce3aa6323c533e3c61cb4f28fe2dd1