Updated openstack/openstack Project: openstack-infra/puppet-openstackid  d24b14fb0e53c6d2b67236bde26c9fb8e28eb45f
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: Ic12edafbb2e33df37c23e3527e961cc71bc67c8c