Updated openstack/openstack Project: openstack-infra/puppet-simpleproxy  f7de5af0f9f6ceae059acd091f6944a94467312e
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I919ae96eda59e10cc11e7392dce3bee21fa270d1