Updated openstack/openstack Project: openstack-infra/puppet-ulimit  5fa98cc809a163b3d580158451168dca5db59649
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I81fe2e21a86be09dbe8b460c0f37573b25993384