Updated openstack/openstack Project: openstack-infra/puppet-ssh  685291c29e288deaaea83815024a0b135827c867
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I37d2467f052556b8490c9bc4ca8d63849b8c09d4