Updated openstack/openstack Project: openstack-infra/puppet-mailman  fde5fe0d6269bf08b97b9e9afe4875d17f60c4f6
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I68cdeb2cb5973f65d375ee4d417cbb1c544eddbf