Updated openstack/openstack Project: openstack-infra/puppet-mediawiki  adec8d34f25e118dd5e5d1858e812e2ca8e9f614
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I41568a6122754097a5d78f0ae0f9b6b61504f551