Updated openstack/openstack Project: openstack-infra/puppet-meetbot  81d84046bb656ea1ee4650bf91a3c7131a9c1bf7
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I26872912f8ff42227f0d88130807f35d900d1209