Updated openstack/openstack Project: openstack-infra/puppet-releasestatus  8613a3f9084aa6c4b3a8f75b3e49b5bbe167a18a
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: If5989b7a19ae2f41d8873ad94b3b0cec532820cc