Updated openstack/openstack Project: openstack-infra/puppet-statusbot  48d4468b0179480f435a124a5e5ea84d22acfc87
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I86231d09496d68df81c92841b3df154026a063a2