Updated openstack/openstack Project: openstack-infra/puppet-snmpd  f6352648ad9cd7fec421e5e4b06f85ffce1dca7c
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I764e7265d5ed281dcfd5ab97b3acb06d81ac9182