Updated openstack/openstack Project: openstack-infra/puppet-reviewday  cb806fc80bf195b4a47aa0b02c7416dca609f175
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I11122f08a677a581941b86b6368600e7c09fa4e2