Updated openstack/openstack Project: openstack-infra/puppet-unattended_upgrades  825b3d68f5bbf64985e500042501dd07512b7f74
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I0fd016546fb86a946f6f484f205b875b2b6f1b9d