Updated openstack/openstack Project: openstack-infra/puppet-openafs  a3ad8d892e4e672c63472601a0e0274e1f90a4bf
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I7c4fd0abf27001ced0ca1670ce574126f8b40cfb