Updated openstack/openstack Project: openstack-infra/puppet-sudoers  ca25acebd5091fc8e91920ca0f1e65fadabe2909
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: Ieb3d56c58b5c63e72d9e9c82db58eeb206e340cf