Updated openstack/openstack Project: openstack-infra/puppet-subunit2sql  ebf675e02cb66c394ce8452a006a9e538e33e496
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: Ic6295486c9ec3c68c6dae90d5d22b01335c10895