Updated openstack/openstack Project: openstack-infra/puppet-packagekit  2897ff76917a9974df84ae10f10d836f29770372
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I5e65feb46bbb6919b9e4479bcfb3ef378b8324ab