Updated openstack/openstack Project: openstack-infra/puppet-planet  6d20d0feb796c13c97ebae0b25152a8fcdddf79c
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I0a2345efec993110c32e4f78d497243e08893376