Updated openstack/openstack Project: openstack-infra/puppet-nodepool  e2b8c279084a768362f95343dc198a0af2bed179
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I17643139de1e79403dce09a5c5d5e66908b5c8d9