Updated openstack/openstack Project: openstack-infra/puppet-mysql_backup  79471eb2203066b7db044479c93b2b8f454e42ed
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: If280787758be4ef18b1346e66756243d306e2406