Updated openstack/openstack Project: openstack-infra/puppet-project_config  231ce9335b11de5d3c9fbfd822696e7c1d12f427
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I2fd13c3aa1b5bb6775c1963f769e6be7e0b8dafb