Updated openstack/openstack Project: openstack-infra/puppet-storyboard  0940bd1e42177035282449d384f8750c0dbb401d
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I0ccda3bc1ee78e42797fd6fcac957f2fdd0c253a