Updated openstack/openstack Project: openstack-infra/puppet-ssl_cert_check  979573cec1d5490a2b13a498b4396a5ea32e8ec3
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I8c8c724b8d45423a3da1b9477f19070a6fb92a05