Updated openstack/openstack Project: openstack-infra/puppet-tmpreaper  c6ca908936a7f3b810703224b8f6c7218deea227
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: Ib4f9a7a66ad4953f826c7010cb49a2f82e6970d6