Updated openstack/openstack Project: openstack-infra/puppet-redis  3c653e53074db5fcd3e99a79d5e9b4009c92f819
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: Icb1243689be7587abc01b6de26da0afcb9a51a7d