Updated openstack/openstack Project: openstack-infra/puppet-unbound  3c4dcf3d5142de45517535ba007c339315b25a83
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I76069001de1266da3232aff95f516b3a213b018d