Updated openstack/openstack Project: openstack-infra/puppet-vcsrepo  1b625245d6313e685e0b2c9922196c803ae82e18
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I11adccf36f9aef32a46de7179d48f6ef296771e2