Updated openstack/openstack Project: openstack-infra/puppet-vinz  305bf73f6e4e31bdd6dbe13d52abab4ac2b4fc82
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: Idc851e3e316fc3d17a9873fc7d9075b2a50f1621