Updated openstack/openstack Project: openstack-infra/puppet-zanata  1619f290d7e6c0d5efe73160d24ab120294ee924
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I5c79b9661db5bec8c2de1763dc96568c9b3bba54