Updated openstack/openstack Project: openstack-infra/puppet-yum  cfc3500aba8127062acb07f267ed276eeafc49e4
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: Id431e76e78b6f550b1bf76eb16e9bdfd407839f4