Updated openstack/openstack Project: openstack-infra/puppet-user  fba7818b6179b23461cd0648666159e8489f500b
Depend on helper gem for spec_helper_acceptance

Instead of keeping a local copy of spec_helper_acceptance.rb and
requiring updates to all modules for any change, we can move it into the
common helper gem and require it from there. This will make it easier to
create and review changes that affect all puppet modules. Also change
the Gemfile to look for the gem in the local workspace if running in a
zuul environment.

Change-Id: I1ce08d6d828822572223b95a77c456759850cd5a