Updated openstack/openstack Project: openstack/glance  68a45793beb6f0f58f804e67731ecee3d743c323
api-ref: update container_format, disk_format

This patch updates the disk_format description to include "ploop"
(added in Ocata).  It also rewrites the container_format description
to indicate that the actual values depend on how the API being queried
is configured.

This patch does not update the api-ref for the deprecated v1 API.

Change-Id: I3385b3f628b74b109c05108e331085c54b0126a1