Updated openstack/openstack Project: openstack/neutron  f311b42d28cdaaa25fab5f9dc4b76d0c294e31dc
Treat Filter as Dict in get_policies

Currently QoS plugin's get_policies doesnt check if the filter
is a dict or not, though similar checks exist in get_rule_types()
and get_policy_rules().
This patch adds the check so that if any consumer passes filters=None
then it is converted to a dict.

TrivialFix
Partial-Bug: #1711165

Change-Id: Id50674c44cd3720dcd8fd6d66361d937599f5992