Updated openstack/openstack Project: openstack/tempest  b42033dd557a5215a88cf9000cfe3c6ac516954a
Remove usage of credentials_factory.AdminManager

We are going to make credentials_factory as stable interfa
and will be removing the AdminManager class which is wrapper
of creating client manager wirth admin cred.

admin manager can be instantiated by providing the admin cred
to clients.Manager.

This commit removes the usage of AdminManager.

Change-Id: I839da90a7f7b5015765a6a67737304ef3109c7c3