Updated openstack/openstack Project: openstack/horizon  ff9cfff507cd2f62d5ce9c2b549ea91a66f4049c
Fix the error format of glance's createImage

Uppon failure, the createImage function returned a raw xhr error
that was incompatible with the wizard modal, making it display
the [Object object] error.

This change makes it so a generic error message is thrown directly
if it's not an handled glance error, so it could be shown on top of
the wizard modal.

Change-Id: I637bf1002630fe21f5ad20a7a047ec65ebc052dc
Closes-bug: #1703708