Updated openstack/openstack Project: openstack/ironic  880d1378cc6bb40bfee57175fc9c375c39d705ac
Add 9.1 to release_mappings

This adds release '9.1' to the list of releases for doing
rolling upgrades.

This also changes the 'pike' release to point to '9.1' since
we are considering 9.1 as the official pike release. In actuality,
it makes no difference since the rpc & object versions are the
same for both 9.0 and 9.1

Change-Id: Id7913a56cd0848a68708a91e6ad055bdee08b465