Updated openstack/openstack Project: openstack/horizon  ffd9140ae8d105e7ac42d1e2e8f35702425f87cc
Make icmp_type/code optional

In Neutron API icmp_type/code are not required fields,
so they should be optional.

Previously we support Nova API which requires icmp_type
and icmp_code but we dropped Nova security group API support in Pike
and we no longer need to care it.

Change-Id: Ie460e763928b4f50f84c38d88556df632078ba42
Closes-Bug: #1712025