Updated openstack/openstack Project: openstack/os-net-config  f8d76d2cdebfa0d06233a59a8f6539207c5b5a4e
Delete controller for standalone OVS bridges

The patch adds an OVS extra parameter to delete controller for bridges
configured with standalone fail mode. By default, bridges are created
without having an openflow controllers. If node is restarted, the bridge
is set to standalone mode but if a service managing the bridge sets a
controller, it will remain in the ovsdb.

As ovs-vswitchd sets the bridge behavior to normal MAC learning switch
only if bridge in standalone mode can't communicate with its controller,
leaving controller defined can cause node outage when bridge is used as
management network. In such case controller service, like
neutron-openvswitch-agent, would need to communicate over management
network but given that bridge is in standalone mode but communicates
with controller, management network won't be reachable. This creates a
chicken-egg problem.

By removing controller by default, ovs-vswitchd implements a normal
action rule to the standalone bridge and service can use the bridge as
management network and eventually set the brdige to secure and set the
flows manually.

See opened Bugzilla for more information:
https://bugzilla.redhat.com/show_bug.cgi?id=1473763

Change-Id: Iad48312667834ea8f5c7145595ae89cb5159b36d