Updated openstack/openstack Project: openstack/ironic  0011a7726cdf2a81d1ffebca18cd74ddc95eccde
Fix names of capabilities for FibreChannel volume boot

Cinder uses 'fibre_channel' to for the volume type for FibreChannel
volumes. This value is returned as 'driver_volume_type' when
initialize_connection API. Then, nova sets the value to a volume target
of ironic.

The capabilities names of interfaces to allow FC boot are expected to
use the value in populating[1]. This patch fixes the capabilities names
from 'fc' to 'fibre_channel'.

[1] https://github.com/openstack/ironic/blob/c0ce6ebf1988c29cc68d16fc2e9142a022f7024b/ironic/drivers/modules/deploy_utils.py#L1263,L1268

Change-Id: I9ff3c55cff035b0ff87b36da820f8129a9ffcbe2
Partial-bug: #1559691